rescaled images
see scale.py for script and instructions.  all images are now 25% their original size.  looks exactly the same, but loads way faster

also now using "jekyll serve --incremental" instead of just "jekyll serve", as this will only build the pages with changes, instead of all of them.  Takes ~2 seconds to generate the updated site, as opposed to ~15 before.  Downside is that it only refreshes the modified pages, so the home page for ex won't refresh this way, which isn't a big deal but good to keep in mind

this has no effect on the published site, just how i view it when making changes